reverted floor to be an integer again

+++ app/src/main/java/com/hotmail/or_dvir/mydoc/ui/new_edit_doctor/NewEditDoctorViewModel.kt
@@ -253,7 +253,7 @@ class NewEditDoctorViewModel(app: Application) : BaseViewModel<NewEditDoctorUiSt
         }
 
         uiState.value?.apply {
-            val floor = if (newInput.isNotBlank()) newInput else null
+            val floor = if (newInput.isNotBlank()) newInput.toInt() else null
             val newAddress =
                 doctor.address?.copy(floor = floor) ?: Address(
                     "",
